few patches

+++ main.py
@@ -2,7 +2,8 @@ import subprocess
 import os
 import shutil
 
-shutil.rmtree('output_images')
+if os.path.exists('output_images'):
+    shutil.rmtree('output_images')
 # Путь к папке с детекторами
 detect_folder = 'detect'
